refactor(popover): move the forecast maths and debug counters out of the view file
UsagePopoverView.swift had grown to 1399 lines and was carrying a SwiftUI view, the pure forecast and colour rules, the shared countdown formatter and a debug-only diagnostic in one file (code review finding #9). The ring colour, run-out, dial-line and spoken-label statics move verbatim into DialForecast.swift as an extension, CountdownFormat moves with them, and the DEBUG body counters move into PopoverBodyCounters.swift. Test references are unchanged and the suite is unchanged at 709.
